% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedGFLMM_fixed_model.R
\name{PedGFLMM_fixed_model}
\alias{PedGFLMM_fixed_model}
\title{PedGFLMM_fixed_model}
\usage{
PedGFLMM_fixed_model(ped, geno, covariate, pos, order, beta_basis, geno_basis,
  base = "bspline", distribution = "binomial", Wald = FALSE)
}
\arguments{
\item{ped}{A data frame containing the pedigree information with the following columns:
\describe{
  \item{ID}{Person ID}
  \item{ped}{pedigree ID,character or numeric allowed.}
  \item{person}{person ID, a unique ID within each pedigree, numeric or character allowed.}
  \item{father}{father ID, NA if no father.}
  \item{mother}{mother ID,  NA if no mother.}
  \item{sex}{sex, coded as 1 for male, 2 for female.}
  \item{trait}{trait phenotype, case-control status coded as 1 for affected and 0 for unaffected. Subjects with missing (NA) will be removed from the analysis.}
}}

\item{geno}{A data frame containing the genotype information.  This is a matrix with genotypes for subjects (rows) at each variant position (columns). The first two columns are required to be named “ped” and “person”, which are used to match subjects to their data in the pedigree data.frame. The genotypes are coded as 0, 1, 2 for autosomal markers (typically a count of the number of the minor alleles).}

\item{covariate}{A data frame containing the covariate information. The first two columns are required to be named “ped” and “person”, which are used to match subjects to their data in the pedigree data frame.}

\item{pos}{Position of the markers in base pairs.}

\item{order}{The order used to generate the B-spline basis.}

\item{beta_basis}{The number of basis functions used to estimate the genetic effect function.}

\item{geno_basis}{The number of basis functions used to estimate the genetic variant functions.}

\item{base}{Can be either 'bspline' or 'fspline'.}

\item{distribution}{This is set to 'binomial'; only binomial is supported.}

\item{Wald}{If Wald is set to true, return the Wald p-value in addition to the LRT p-value (Default: Wald = FALSE).}
}
\value{
A list containing the following components:
\describe{
  \item{LRT}{The p-value based on a likelihood ratio test}
  \item{Wald}{The p-value based on a Wald test, returned if 'Wald' is TRUE}
}
}
\description{
Computes the PedGFLMM statistics under a fixed model.
}
\examples{
data(exampleData)

betabasis_Bsp = 10
genobasis_Bsp = 10

betabasis_Fsp = 11
genobasis_Fsp = 11
order  =   4

fixed_bsp=PedGFLMM_fixed_model(ped = Ped, geno = as.matrix(geno), pos = snpPos$pos,
    order = order, beta_basis=betabasis_Bsp, geno_basis = genobasis_Bsp,
    covariate = as.matrix(cov), base = "bspline", distribution = 'binomial')
fixed_bsp

fixed_fsp=PedGFLMM_fixed_model(ped = Ped, geno = as.matrix(geno), pos = snpPos$pos,
    order = order, beta_basis=betabasis_Fsp, geno_basis = genobasis_Fsp,
    covariate = as.matrix(cov), base = "fspline", distribution = 'binomial')
fixed_fsp
}
\references{
Chiu CY, Yuan F, Zhang BS, Yuan A, Li X, Fang HB, Lange K, Weeks DE, Wilson AF, Bailey-Wilson JE, Lakhal-Chaieb ML, Cook RJ, McMahon FJ, Amos CI, Xiong MM, and Fan RZ (2019) Pedigree-based linear mixed models for association analysis of quantitative traits with next-generation sequencing data. Genetic Epidemiology 43(2):189-206.

Fan RZ, Wang YF, Mills JL, Wilson AF, Bailey-Wilson JE, and Xiong MM (2013) Functional linear models for association analysis of quantitative traits. Genetic Epidemiology 37 (7):726- 742.

Fan RZ, Wang YF, Mills JL, Carter TC, Lobach I, Wilson AF, Bailey-Wilson JE, Weeks DE, and Xiong MM (2014) Generalized functional linear models for case-control association studies. Genetic Epidemiology 38 (7):622-637.

Jiang YD, Chiu CY, Yan Q, Chen W, Gorin MB, Conley YP, Lakhal-Chaieb ML, Cook RJ, Amos CI, Wilson AF, Bailey-Wilson JE, McMahon FJ, Vazquez AI, Yuan A, Zhong XG, Xiong MM, Weeks DE, and Fan RZ (2019) Gene-based association testing of dichotomous traits with generalized linear mixed models for family data.

Schaid DJ, McDonnell SK, Sinnwell JP, and Thibodeau SN (2013) Multiple genetic variant association testing by collapsing and kernel methods with pedigree or population structured data. Genetic Epidemiology 37:409-418.
}
\seealso{
\code{\link{PedGFLMM_beta_smooth_only}}, \code{\link{PedGLMM_additive_effect_model}}, \code{\link{exampleData}}
}
